:recycle: refactory: adicionar funções compartilhadas para entrada, saída e conversão
