+++ src/shared.h
@@ -1,0 +1,9 @@
+#ifndef SHARED_H
+#define SHARED_H
+
+// Declarações das funções
+double readInputValue(const char* message);
+void printResult(double value1, double value2, const char* unit1, const char* unit2);
+double convert(double value, double factor);
+
+#endif
